cleaned up protected var

+++ src/ESPRotary.cpp
@@ -9,7 +9,7 @@
 
 /////////////////////////////////////////////////////////////////
 
-ESPRotary::ESPRotary(int pin1, int pin2, int steps_per_click /* = 1 */,  int lower_bound /* = INT16_MIN */, int upper_bound /* = INT16_MAX */, int inital_pos /* = 0 */) {
+ESPRotary::ESPRotary(byte pin1, byte pin2, byte steps_per_click /* = 1 */,  int lower_bound /* = INT16_MIN */, int upper_bound /* = INT16_MAX */, int inital_pos /* = 0 */) {
   this->pin1 = pin1;
   this->pin2 = pin2;
   this->lower_bound = (lower_bound < upper_bound) ? lower_bound : upper_bound;
